requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gitpython==3.1.40
tree-sitter==0.20.4
typer==0.9.0
rich==13.7.0
requests==2.31.0
openai==1.3.7
pydantic==2.5.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5 